feat: merge crates untrusted and syscall into sallyport
This is in preparation to split-out sallyport in its own git repo and
official crate.

The sallyport crate gathers a link-time symbol preventing multiple
sallyport versions to be used in one executable.

The loader now checks via the sallyport symbol, if the shim
follows the sallyport ABI the loader supports.

Signed-off-by: Harald Hoyer <harald@redhat.com>